Need cast to bigint

+++ src/variable.cpp
@@ -3417,7 +3417,7 @@ void Variable::particle_vector(char *word, Tree **tree,
     newtree->carray = (char *) &species[0].magmoment;
   }
 
-  if (particle->nlocal*newtree->nstride > MAXSMALLINT)
+  if ((bigint)particle->nlocal*newtree->nstride > MAXSMALLINT)
     error->all(FLERR,"Too many particles per processor for particle-style variable");
 }
 
@@ -3474,7 +3474,7 @@ void Variable::grid_vector(char *word, Tree **tree,
   else if (strcmp(word,"czhi") == 0)
     newtree->carray = (char *) &cells[0].hi[2];
 
-  if (grid->nlocal*newtree->nstride > MAXSMALLINT)
+  if ((bigint)grid->nlocal*newtree->nstride > MAXSMALLINT)
     error->all(FLERR,"Too many grid cells per processor for grid-style variable");
 }
